Revert "Split into multiple file, so it is possible to reuse the logic in a script"
